Move the location of distance.hpp
